kwsys_common: Implement nightly testing for stage-based workflow
KWSys development is moving to a Fork & Merge Request model where we
maintain a "stage" for each target branch (e.g. master) that consists of
the target branch plus some set of MRs merged on top of it.  The job of
this testing script is now to test one of the integration branches or
one of their stage branches.  Add infrastructure to create or convert
the local test repository such that it has tracking branches for each
stage branch.  Select one to check out and follow for testing.